added test for Andronescu 2007 params
